git-svn-id: svn+ssh://svn.fcdonders.nl/home/svnroot/fieldtrip/trunk@3238 0cf7c7f0-3615-4144-b4e6-68da3bce3cd0

+++ ft_resampledata.m
@@ -142,7 +142,7 @@ if usefsample && usetime
 end
 
 % remember the original sampling frequency in the configuration
-cfg.origfs = data.fsample;
+cfg.origfs = double(data.fsample);
 
 if usefsample
   %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
